Clean up math package
